feat: create navigation system using reduxjs/toolkit
This commit creates a component as Link, which accepts three parameters as
'to', 'children', and '...rest'. First one being the path name to intended
page, the second one is title of the link and the third one is extra prop
optionality. This Link component takes the path as the first prop and updates
the window history as well as the redux store state for navigation (ie. state.navigation.activePage)