fix(Link comp): fix button tag overlays
As button tags can't be decendent of another button tags, the button tag in Link
component is making it errorious to use it to wrap our Button component. Therefore
as the button tag in Link component also does not has any essence, this commit removes
that button tag and passes the 'children' prop directly inside the anchor tags. A div
can be added in future if any other problem occurs