feat(new component): create TextThumbnail for handling no image blogs
This commit creates a new component named TextThumbnail, which can is very similar
in nature with the RectImage component. This component takes 6 props and provides
a custom shaped div with a custom text inside. This can be used in any kind of
thumbnails, with no image and only a text. The 6 props that this component accepts
are - title, width, height, dataUnit, aspectRatio, background.
title being the text to display in the box. width and height can be provided for
custom sizing or only one can be provided with an object with two numbers being
desired aspect ratio of the box to the aspectRatio prop. The width and/or height
props expect only a number and the unit should be provided through the dataUnit
prop. The background prop expects a hex code or any other css color value. This
color will be applied as the background of the box.