feat(landing/projects): use json instead of js for projectsData
This commit finally resolves the issue that I was going through. I finally
found a way to send/store html/jsx expressions in json and use that in the
react components. I just stored the jsx tags inside a string. This commit
now uses the dangeriouslySetInnerHTML special prop to render our json jsx
inside that div.