feat(landing/projects): create functionalities for making projects sec responsive
This commit makes the projects section in the landing page responsive. In addition to that
this commit also creates functionality to hide the preview panel, when the website width
goes to less than 800px. This commit also changes the behaviour of the projects cards in
projects-sec-cards-panel to navigate to the particular projects page, rather than changing
the activeProject peice of state. In conclusion the activeProject state is of no use when
the screen size is less than 800px