feat(navbar): create a 3 way responsive mechanism
This commit removes the boolean state responsible for detecting if the screen width
is smaller than 1100px or not, and replaces it with another state called screenWidthSts
which will store either of 3 string values, 'alpha', 'beta', 'gamma'. The state will
be updated through the event listeners. When the width will be more than 1100, the state
will be 'alpha', for more than 600 but less than 1100 will be 'beta' and less than 600px
will be gamma. Depending upon which value that state has, the component returns a chunk
of jsx which are specifically made to go with that width.