fix(navbar): fix render of pages on navigation back & forward
Due to some reasons the jsx chunk in the renderNavBtns const has to be rendered in the navbar
in order to the pages to be rendered automaticly when coming back to pages from any other page
using the back btns of the browser. I still did not figured out why, but this is just happening.
Therefore this commit creates a div with classname 'hidden', which hids the div and calls that
const inside that div. That means in the background the nav buttons are now being rendered
parmanently in the beta and gamma navbars.