cleanup(landingpage): delete section files & empty folders
as sections are already migrated to be widgets inside src/components/widgets directory, the files inside
the sections directory in landingpage are unused and unneccessary, therefore this commit will delete those
files and the sections directory. This commit will also delete the elements directory, as it is also useless,
without the section files, and the data directory, as the data files are already migrated to server and deleted
from this directory. In total this commit deletes three directories as, elements, data & sections