feat(project panels): use uniqueKey from api json for better data match
The project section is getting data from the backend, but the preview panels are being built
in the frontend. Therefore there's a need to add better validation for data mathing among these
two data sources. The backend repo is already introduced with a new property in the projectsWidget
dataset as uniqueKey. This property is a unique string with only letters, adopted from the title
itself. This commit changes the validation process of PanelRenderer to use that uniqueKey prop
from the backend json